requirements.txt
appdirs
click
path.py
pyotp
toml # TODO: move to tomli
cryptography
keyring
textual
clipman
typing_extensions
